src: fix FatalException deprecation warning
The non-isolate version of node::FatalException() has been deprecated
in io.js; more precisely, it's deprecated in node.js v0.12 as well but
a bug in joyent/node prevents the deprecation warning from printing.

Switch to NanMakeCallback() because we don't have to catch exceptions
ourselves.  Exceptions simply unwind the stack until they encounter a
try/catch block is encountered or the fatal exception handler in the
io.js or node.js run-time.

Fixes: https://github.com/bnoordhuis/node-unix-dgram/issues/23