src: return send error in callback if present
Otherwise emit an `error` event.

This behaviour emulates the behaviour of `udp` core module:
"If an error occurs and a callback is given, the error will be passed as
the first argument to the callback. If a callback is not given, the
error is emitted as an 'error' event on the socket object."